# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AraLink API",
    "version": "1.0.0",
    "description": "AraLink Translation API — ترجم أي رابط (يوتيوب، مقالات، مواقع) إلى أي لغة.\n\nجميع نقاط النهاية سلمت تحت `/api`. نقاط الإدارة تتطلب ترويسة `x-admin-token` (انظر ADMIN_TOKEN في .env.example)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "رمز الإدارة ADMIN_TOKEN — يفتح مسارات /api/settings و /api/stats وإحصاءات /api/jobs"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "رمز خطأ قصير (مثل video-not-found)" },
          "message": { "type": "string", "description": "وصف عربي قابل للعرض" },
          "code": { "type": "string" }
        }
      },
...
      "JobStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["queued", "running", "done", "failed", "cancelled"] },
          "progress": { "type": "number", "minimum": 0, "maximum": 100 },
          "result": { "type": "object", "additionalProperties": true },
          "error": { "type": "string" }
        }
      }
    },
    "parameters": {
 "AdminHeader": {
   "name": "x-admin-token",
   "in": "header",
   "required": true,
   "schema": {
    "type": "string"
   }
  },
  "ProjectHeader": {
   "name": "x-project-token",
   "in": "header",
   "required": false,
   "description": "توكن مالك المشروع (يُعاد مرة واحدة عند POST /api/projects) — مطلوب لحذف وظيفة مرتبطة بمشروع",
   "schema": {
    "type": "string"
   }
  }
    }
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": ["jobs"],
        "summary": "إحصاءات الطابور (عدّادات إجمالية — للأدمن)",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AdminHeader" }],
        "responses": {
          "200": { "description": "عدّادات {concurrency, running, queued, tracked, maxQueued}" },
          "401": { "description": "رمز إدارة خاطئ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "503": { "description": "ADMIN_TOKEN غير مضبوط — المسار معطّل", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "tags": ["jobs"],
        "summary": "حالة وظيفة (رابط قدرة UUID — بلا مصادقة عمدًا، EventSource لا يرسل ترويسات)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "الحالة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatus" } } } },
          "404": { "description": "غير موجودة أو انتهت صلاحيتها", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "tags": ["jobs"],
        "summary": "إلغاء وظيفة قيد التنفيذ (وظائف المشاريع تتطلّب توكن المالك أو الأدمن)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }, { "$ref": "#/components/parameters/ProjectHeader" }],
        "responses": {
          "200": { "description": "أُلغيت أو كانت منتهية أصلًا" },
          "404": { "description": "غير موجودة أو لست مالك مشروعها", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/jobs/{id}/stream": {
      "get": {
        "tags": ["jobs"],
        "summary": "بث تقدم وظيفة (SSE)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "تدفق أحداث SSE", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
          "404": { "description": "غير موجودة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}